+++ src/math.rs
@@ -1,11 +1,9 @@
-use crate::Float;
-
 pub mod statistics;
 pub mod vector;
 
 #[must_use]
 #[inline]
-pub fn logistic(x: Float) -> Float {
+pub fn logistic(x: f64) -> f64 {
     1.0 / (1.0 + (-x).exp())
 }
