+++ src/web/routes/player.rs
@@ -27,7 +27,6 @@ use crate::web::partials::*;
 use crate::web::response::CustomResponse;
 use crate::web::routes::player::partials::*;
 use crate::web::TrackingCode;
-use crate::Float;
 
 pub mod partials;
 
@@ -358,7 +357,7 @@ pub async fn get(
                                                 div.level-item.has-text-centered {
                                                     div {
                                                         p.heading { "В час" }
-                                                        p.title { (render_float(stats_delta.frags as Float / battle_life_time as Float * 3600.0, 1)) }
+                                                        p.title { (render_float(stats_delta.frags as f64 / battle_life_time as f64 * 3600.0, 1)) }
                                                     }
                                                 }
                                             }
@@ -406,7 +405,7 @@ pub async fn get(
                                                 div.level-item.has-text-centered {
                                                     div {
                                                         p.heading { "В час" }
-                                                        p.title { (render_float(stats_delta.wins as Float / battle_life_time as Float * 3600.0, 1)) }
+                                                        p.title { (render_float(stats_delta.wins as f64 / battle_life_time as f64 * 3600.0, 1)) }
                                                     }
                                                 }
                                             }
@@ -545,14 +544,14 @@ async fn make_predictions(
     redis: &mut MultiplexedConnection,
     account_id: i32,
     tanks: &[Tank],
-) -> crate::Result<IndexMap<i32, Float>> {
+) -> crate::Result<IndexMap<i32, f64>> {
     let account_factors = match get_account_factors(redis, account_id).await? {
         Some(factors) => factors,
         None => return Ok(IndexMap::new()),
     };
     let vehicles_factors = get_all_vehicle_factors(redis).await?;
 
-    let mut predictions: IndexMap<i32, Float> = tanks
+    let mut predictions: IndexMap<i32, f64> = tanks
         .iter()
         .map(|tank| remap_tank_id(tank.statistics.base.tank_id))
         .filter_map(|tank_id| {
@@ -569,11 +568,11 @@ async fn make_predictions(
 }
 
 fn top_tanks_column(
-    predictions: &IndexMap<i32, Float>,
+    predictions: &IndexMap<i32, f64>,
     predicate: fn(&Vehicle) -> bool,
     title: &str,
 ) -> Markup {
-    let entries: Vec<(&i32, &Float)> = predictions
+    let entries: Vec<(&i32, &f64)> = predictions
         .iter()
         .filter(|(tank_id, _)| predicate(&get_vehicle(**tank_id)))
         .take(5)
